+++ ggml/src/ggml-openvino/ggml-openvino-op-support.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "ggml-backend.h"
+#include "ggml.h"
+
+bool ggml_openvino_device_supports_op_impl(ggml_backend_dev_t dev, const ggml_tensor * op);
