+++ ggml/src/ggml-openvino/ggml-openvino.cpp
@@ -4,7 +4,7 @@
 #include "ggml-backend.h"
 #include "ggml-impl.h"
 #include "ggml-openvino-extra.h"
-#include "ggml-openvino/openvino/op_table.h"
+#include "ggml-openvino-op-support.h"
 #include "ggml-openvino/utils.h"
 #include "ggml-quants.h"
 #include "ggml.h"
@@ -21,7 +21,6 @@
 #include <openvino/runtime/intel_gpu/ocl/ocl.hpp>
 #include <openvino/runtime/intel_npu/level_zero/level_zero.hpp>
 #include <openvino/runtime/tensor.hpp>
-#include <set>
 #include <string>
 #include <vector>
 
@@ -882,546 +881,8 @@ static ggml_backend_buffer_type_t ggml_backend_openvino_device_get_host_buffer_t
     return ggml_backend_openvino_host_buffer_type(ctx->device);
 }
 
-static bool has_view_op_input(const ggml_tensor * op) {
-    for (int i = 0; i < GGML_MAX_SRC; i++) {
-        if (op->src[i] == nullptr) {
-            break;
-        }
-        if (op->src[i]->op == GGML_OP_VIEW) {
-            return true;
-        }
-    }
-    return false;
-}
-
-static bool has_non_contiguous_view_input(const ggml_tensor * op) {
-    for (int i = 0; i < GGML_MAX_SRC; i++) {
-        if (op->src[i] == nullptr) {
-            break;
-        }
-        if (op->src[i]->op == GGML_OP_VIEW && !ggml_is_contiguous(op->src[i])) {
-            return true;
-        }
-    }
-    return false;
-}
-
-static bool is_supported_flash_attn_pattern(const ggml_tensor * op) {
-    // pattern of q,k,v should be q->op==PERMUTE, q->src[0]->op==VIEW, q->src[0]->src[0]->view_src==nullptr
-    for (int i = 0; i < 3; i++) {
-        const ggml_tensor * src = op->src[i];
-        if (src->op != GGML_OP_PERMUTE || src->src[0] == nullptr || src->src[0]->op != GGML_OP_VIEW ||
-            src->src[0]->src[0] == nullptr || src->src[0]->src[0]->view_src != nullptr) {
-            return false;
-        }
-    }
-    return true;
-}
-
-static bool is_gemma3n_flash_attn_pattern(const ggml_tensor * op) {
-    if (!is_supported_flash_attn_pattern(op)) {
-        return false;
-    }
-
-    const ggml_tensor * q_base =
-        op->src[0] != nullptr && op->src[0]->src[0] != nullptr ? op->src[0]->src[0]->src[0] : nullptr;
-    const ggml_tensor * k_base =
-        op->src[1] != nullptr && op->src[1]->src[0] != nullptr ? op->src[1]->src[0]->src[0] : nullptr;
-    const ggml_tensor * v_base =
-        op->src[2] != nullptr && op->src[2]->src[0] != nullptr ? op->src[2]->src[0]->src[0] : nullptr;
-
-    if (q_base == nullptr || q_base->op != GGML_OP_ROPE) {
-        return false;
-    }
-
-    // gemma3n direct attention path (no KV cache): q=ROPE, k=ROPE, v=RMS_NORM
-    // Only match this specific pattern to avoid falsely catching other models
-    // (e.g. Gemma4) that also use scale=1.0 with KV-cache backed attention.
-    const bool is_qkv_direct =
-        k_base != nullptr && v_base != nullptr && k_base->op == GGML_OP_ROPE && v_base->op == GGML_OP_RMS_NORM;
-
-    return is_qkv_direct;
-}
-
-static bool checked_mul_size(size_t a, size_t b, size_t & out) {
-    if (a == 0 || b == 0) {
-        out = 0;
-        return true;
-    }
-    if (a > SIZE_MAX / b) {
-        return false;
-    }
-    out = a * b;
-    return true;
-}
-
-static bool tensor_view_fits_src_buffer(const ggml_tensor * tensor) {
-    if (tensor->view_src == nullptr) {
-        return true;
-    }
-
-    const size_t src_nbytes = ggml_nbytes(tensor->view_src);
-    if (tensor->view_offs > src_nbytes) {
-        return false;
-    }
-
-    const size_t tensor_nbytes = ggml_nbytes(tensor);
-    return tensor_nbytes <= src_nbytes - tensor->view_offs;
-}
-
-static bool cpy_output_view_is_supported(const ggml_tensor * op) {
-    if (op->view_src == nullptr) {
-        return true;
-    }
-
-    if (!tensor_view_fits_src_buffer(op)) {
-        return false;
-    }
-
-    return ggml_nbytes(op) == 0 || ggml_is_contiguous(op);
-}
-
-static bool mul_mat_id_requires_large_tmp(const ggml_tensor * op) {
-    const ggml_tensor * as = op->src[0];
-    const ggml_tensor * ids = op->src[2];
-    if (as == nullptr || ids == nullptr) {
-        return true;
-    }
-
-    // The MXFP4 MUL_MAT_ID translation (translate_mul_mat_id_mxfp4_packed in mul_mat_id.cpp)
-    // materializes selected expert weights with shape [n_tokens, n_used, rows, k]. Skip cases that
-    // would create a very large temporary and let the scheduler fall back instead. Every other weight
-    // type goes through GatherMatmul, which never materializes this temporary.
-    size_t tmp_elems = 1;
-    if (!checked_mul_size(tmp_elems, static_cast<size_t>(ids->ne[1]), tmp_elems) ||
-        !checked_mul_size(tmp_elems, static_cast<size_t>(ids->ne[0]), tmp_elems) ||
-        !checked_mul_size(tmp_elems, static_cast<size_t>(as->ne[1]), tmp_elems) ||
-        !checked_mul_size(tmp_elems, static_cast<size_t>(as->ne[0]), tmp_elems)) {
-        return true;
-    }
-
-    size_t tmp_bytes = 0;
-    if (!checked_mul_size(tmp_elems, sizeof(float), tmp_bytes)) {
-        return true;
-    }
-
-    static constexpr size_t mul_mat_id_tmp_limit = 1ULL << 30;  // 1 GiB
-    return tmp_bytes > mul_mat_id_tmp_limit;
-}
-
-static bool tensor_name_starts_with(const ggml_tensor * tensor, const char * prefix) {
-    return tensor != nullptr && strncmp(tensor->name, prefix, strlen(prefix)) == 0;
-}
-
-static bool is_msa_block_mask_expansion(const ggml_tensor * op) {
-    if (tensor_name_starts_with(op, "msa_")) {
-        return true;
-    }
-
-    const ggml_tensor * src = op->src[0];
-    while (src != nullptr && (src->op == GGML_OP_RESHAPE || src->op == GGML_OP_REPEAT)) {
-        if (tensor_name_starts_with(src, "msa_block_mask")) {
-            return true;
-        }
-        src = src->src[0];
-    }
-
-    return tensor_name_starts_with(src, "msa_block_mask");
-}
-
-static bool is_op_unsupported_case(const ggml_tensor * op) {
-    if (is_msa_block_mask_expansion(op)) {
-        return true;
-    }
-
-    switch (op->op) {
-    case GGML_OP_CONCAT: {
-        if (op->type == GGML_TYPE_I64) {
-            return true;
-        }
-        if (ggml_openvino_get_device_name() == "GPU" && op->type == GGML_TYPE_BF16 && has_view_op_input(op)) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_SET: {
-        const auto nb1 = static_cast<size_t>(op->op_params[0]);
-        const auto nb2 = static_cast<size_t>(op->op_params[1]);
-        const auto nb3 = static_cast<size_t>(op->op_params[2]);
-
-        // OpenVINO SET translation currently supports dst layouts that match src0 strides.
-        if (op->src[0] == nullptr || nb1 != op->src[0]->nb[1] || nb2 != op->src[0]->nb[2] || nb3 != op->src[0]->nb[3]) {
-            // std::cout << "Unsupported SET op with dst nb1=" << nb1 << ", nb2=" << nb2 << ", nb3=" << nb3
-            //           << " that does not match src0 strides nb[1]="
-            //           << (op->src[0] != nullptr ? std::to_string(op->src[0]->nb[1]) : "null")
-            //           << ", nb[2]=" << (op->src[0] != nullptr ? std::to_string(op->src[0]->nb[2]) : "null")
-            //           << ", nb[3]=" << (op->src[0] != nullptr ? std::to_string(op->src[0]->nb[3]) : "null")
-            //           << std::endl;
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_GET_ROWS:
-    case GGML_OP_SET_ROWS: {
-        if (op->ne[3] != 1) {
-            return true;
-        }
-        if (op->op == GGML_OP_GET_ROWS && ggml_openvino_get_device_name() == "GPU" &&
-            op->src[0]->type == GGML_TYPE_BF16) {
-            return true;
-        }
-        if (op->ne[0] == 256 && (op->src[0]->type == GGML_TYPE_Q4_K || op->src[0]->type == GGML_TYPE_Q5_K ||
-                                 op->src[0]->type == GGML_TYPE_Q4_1 || op->src[0]->type == GGML_TYPE_Q5_1)) {
-            // These are all f16-arithmetic dequant rounding errors that intermittently exceed the
-            // tight 1e-7 NMSE threshold depending on the random test data (see ggml-quants.cpp
-            // make_int8_weights/make_int4_weights: dequant is done in f16, not f32, to keep the
-            // Convert/Subtract/Multiply chain fusable into GatherMatmulCompressed/FullyConnectedCompressed
-            // for the shared non-test code paths).
-            return true;
-        }
-
-        break;
-    }
-    case GGML_OP_RESHAPE: {
-        if (strncmp(op->name, "ffn_norm_exps", sizeof("ffn_norm_exps") - 1) == 0) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_ADD:
-    case GGML_OP_MUL:
-    case GGML_OP_SUB: {
-        if (op->src[1]->op == GGML_OP_PERMUTE) {
-            return true;
-        }
-        for (int i = 0; i < 4; i++) {
-            if (op->src[0]->ne[i] != op->src[1]->ne[i] && (op->src[0]->ne[i] != 1 && op->src[1]->ne[i] != 1)) {
-                return true;
-            }
-        }
-        break;
-    }
-    case GGML_OP_ADD_ID: {
-        // Keep support aligned with the CPU backend implementation, which only handles f32 inputs/output and i32 ids.
-        if (op->type != GGML_TYPE_F32 || op->src[0]->type != GGML_TYPE_F32 || op->src[1]->type != GGML_TYPE_F32 ||
-            op->src[2]->type != GGML_TYPE_I32) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_DIV: {
-        // The GPU plugin can fuse broadcast DIV into the preceding FFN GEMM path
-        // and produce infs for per-channel scale vectors. Keep those DIVs on CPU
-        // until the fused GPU kernel is reliable. (falied case llama-arch-test mpt)
-        if (ggml_openvino_get_device_name() == "GPU" && op->src[1]->ne[0] == op->ne[0] &&
-            op->src[1]->ne[1] == 1 && op->src[1]->ne[2] == 1 && op->src[1]->ne[3] == 1) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_SUM_ROWS: {
-        // if the input is PERMUTE skip
-        if (op->src[0]->op == GGML_OP_PERMUTE) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_FLASH_ATTN_EXT: {
-        float scale = 1.0f;
-        float max_bias = 0.0f;
-        float logit_softcap = 0.0f;
-        const auto * op_params = op->op_params;
-        memcpy(&scale, (const float *) op_params + 0, sizeof(float));
-        memcpy(&max_bias, (const float *) op_params + 1, sizeof(float));
-        memcpy(&logit_softcap, (const float *) op_params + 2, sizeof(float));
-
-        // Keep gemma3n flash-attn pattern on CPU for GPU runs to avoid
-        // accuracy drift in the OpenVINO path. Restrict by scale=1.0 to avoid
-        // affecting non-gemma3n models such as Llama-3.2.
-        if (fabsf(scale - 1.0f) < 1e-6f && is_gemma3n_flash_attn_pattern(op)) {
-            return true;
-        }
-
-        if (op->src[4] != nullptr) {
-            // GGML_LOG_WARN("OpenVINO backend does not support FLASH_ATTN_EXT with sinks\n");
-            return true;
-        }
-        if (!is_supported_flash_attn_pattern(op)) {
-            return true;
-        }
-        if (max_bias > 0) {
-            // GGML_LOG_WARN("OpenVINO backend does not support FLASH_ATTN_EXT with max_bias > 0\n");
-            return true;
-        }
-        if (logit_softcap != 0) {
-            // GGML_LOG_WARN("OpenVINO backend does not support FLASH_ATTN_EXT with logit_softcap != 0\n");
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_PERMUTE: {
-        if (op->type == GGML_TYPE_BF16) {
-            // err msg: [GPU] Could not find a suitable kernel for transpose
-            // GGML_LOG_WARN("OpenVINO backend does not support PERMUTE with BF16 type\n");
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_CPY: {
-        if (op->src[0]->type == GGML_TYPE_BF16 || op->src[1]->type == GGML_TYPE_BF16) {
-            // GGML_LOG_WARN("OpenVINO backend does not support CPY with non-contiguous data or bf16 types\n");
-            return true;
-        }
-        if (ggml_nelements(op->src[0]) != ggml_nelements(op->src[1])) {
-            return true;
-        }
-        // op test case with non-contiguous src or dst
-        if ((op->ne[0] == 3 && op->ne[1] == 4 && op->ne[2] == 3 && op->ne[3] == 2) ||
-            (op->ne[0] == 1 && op->ne[1] == 4 && op->ne[2] == 3 && op->ne[3] == 2) ||
-            (op->ne[0] == 2 && op->ne[1] == 4 && op->ne[2] == 3 && op->ne[3] == 2)) {
-            return true;
-        }
-        if (!cpy_output_view_is_supported(op)) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_MUL_MAT: {
-        if (ggml_openvino_get_device_name() == "GPU" && op->src[0] != nullptr && op->src[1] != nullptr &&
-            ggml_is_quantized(op->src[0]->type) && strcmp(op->src[0]->name, "a") == 0 &&
-            strcmp(op->src[1]->name, "b") == 0 && op->src[0]->ne[1] == 1 && op->src[1]->ne[1] == 64 &&
-            op->src[0]->ne[0] == 256 && op->src[1]->ne[0] == 256) {
-            return true;
-        }
-        if (op->src[0]->ne[3] != op->src[1]->ne[3] && op->src[0]->ne[3] != 1 && op->src[1]->ne[3] != 1) {
-            return true;
-        }
-        if (op->src[0]->op == GGML_OP_VIEW && op->src[1]->op == GGML_OP_VIEW) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_MUL_MAT_ID: {
-        // Single-expert (or empty) MUL_MAT_ID is a degenerate shape that stresses GatherMatmul edge
-        // cases and never occurs in real MoE; let it fall back to CPU.
-        if (op->src[0] != nullptr && op->src[0]->ne[2] <= 1) {
-            return true;
-        }
-        if (ggml_openvino_get_device_name() == "GPU" && op->src[0] != nullptr && op->src[0]->type == GGML_TYPE_BF16) {
-            return true;
-        }
-        // GPU MUL_MAT_ID uses a Gather+MatMul fallback because the GPU plugin rejects internal
-        // GatherMatmul for these test shapes. Skip cases that would materialize a large selected
-        // expert-weight temporary.
-        if (ggml_openvino_get_device_name() == "GPU" && mul_mat_id_requires_large_tmp(op)) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_ROPE: {
-        const int32_t * op_params = op->op_params;
-        const int n_dims = op_params[1];
-        const int mode = op_params[2];
-        if (mode != GGML_ROPE_TYPE_NORMAL && mode != GGML_ROPE_TYPE_NEOX && mode != GGML_ROPE_TYPE_IMROPE) {
-            // GGML_LOG_WARN("OpenVINO backend does not support ROPE with mode %d\n", mode);
-            return true;
-        }
-        const int64_t head_dim = op->src[0]->ne[0];
-        const int64_t rope_dims = n_dims == 0 ? head_dim : n_dims;
-        if (rope_dims <= 0 || rope_dims > head_dim || (rope_dims % 2) != 0) {
-            // GGML_LOG_WARN("OpenVINO backend does not support ROPE with n_dims %d and src[0]->ne[0] %ld\n", n_dims,
-            //               op->src[0]->ne[0]);
-            return true;
-        }
-        if (op->type != GGML_TYPE_F32 && op->type != GGML_TYPE_F16) {
-            // GGML_LOG_WARN("OpenVINO backend does not support ROPE with type %s\n", ggml_type_name(op->type));
-            return true;
-        }
-        if (op->src[0]->op == GGML_OP_VIEW) {
-            if (op->src[0]->view_src->ne[1] != op->src[0]->ne[2]) {
-                // GGML_LOG_WARN(
-                //     "OpenVINO backend does not support ROPE with src[0]->view_src->ne[1] %ld != src[0]->ne[2] "
-                //     "%ld\n",
-                //     op->src[0]->view_src->ne[1], op->src[0]->ne[2]);
-                return true;
-            }
-        }
-        if (mode == GGML_ROPE_TYPE_IMROPE &&
-            (op->src[2] != 0 || ((const float *) op_params)[6] != 1 || ((const float *) op_params)[7] != 0 ||
-             ((const float *) op_params)[8] != 1)) {
-            // GGML_LOG_WARN("OpenVINO backend does not support IMROPE with freq_factors, freq_scale, ext_factor, and attn_factor\n");
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_TRANSPOSE: {
-        // if the type is bf16, will return true
-        if (op->type == GGML_TYPE_BF16) {
-            // GGML_LOG_WARN("OpenVINO backend does not support CONT with BF16 type\n");
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_REPEAT: {
-        if (ggml_openvino_get_device_name() == "GPU" && op->type == GGML_TYPE_BF16) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_GATED_DELTA_NET: {
-        // enable after https://github.com/openvinotoolkit/openvino/pull/35917 is included in OV release
-        // return true;
-        // if (ggml_openvino_get_device_name() == "GPU" && op->src[0]->ne[2] > 1) {
-        //     // CVS-186471
-        //     return true;
-        // }
-        if (op->src[2]->op == GGML_OP_PERMUTE) {
-            return true;
-        }
-        // kda (per-key-dimension gating) not supported by fused GatedDeltaNet op
-        if (op->src[3]->ne[0] != 1) {
-            return true;
-        }
-        // K > 1 (multiple state snapshots) not supported by fused op
-        if (((const int32_t *) op->op_params)[0] > 1) {
-            return true;
-        }
-        break;
-    }
-    case GGML_OP_SSM_CONV: {
-        // qwen3next is numerically unstable with OpenVINO SSM_CONV.
-        // Keep this op on CPU until the OpenVINO implementation is fixed.
-        // return true;
-        break;
-    }
-    case GGML_OP_VIEW: {
-        // Skip TOPK_MOE fused tests until it is fully supported.
-        // The argsort_top_k VIEW wrapping ARGSORT is named "selected_experts" in test_topk_moe.
-        if (strcmp(op->name, "selected_experts") == 0) {
-            return true;
-        }
-        break;
-    }
-    default:
-        break;
-    }
-    return false;
-}
-
 static bool ggml_backend_openvino_device_supports_op(ggml_backend_dev_t dev, const ggml_tensor * op) {
-    GGML_ASSERT(dev->reg != nullptr);
-
-    static std::unordered_set<ggml_type> supported_types{
-        GGML_TYPE_F32,  GGML_TYPE_F16,  GGML_TYPE_BF16, GGML_TYPE_I64,  GGML_TYPE_I32,  GGML_TYPE_Q4_0,
-        GGML_TYPE_Q4_1, GGML_TYPE_Q4_K, GGML_TYPE_Q5_1, GGML_TYPE_Q5_K, GGML_TYPE_Q8_0, GGML_TYPE_Q6_K,
-        GGML_TYPE_MXFP4};
-
-    // derive supported op sets from the op_table map, keys in
-    // the map use the full macro name (e.g. "GGML_OP_ADD"), while
-    // the ggml_*_op_name() helpers return only the trailing part (e.g. "ADD").
-    // each set is built once and cached.
-    static const auto build_supported_sets = [] {
-        const auto & table = ov::frontend::ggml::get_supported_ops();
-        std::unordered_set<ggml_op> ops;
-        std::unordered_set<ggml_unary_op> unary_ops;
-        std::unordered_set<ggml_glu_op> glu_ops;
-
-        // GGML_OP_NONE has no translator but is always safe to add to the supported set.
-        ops.insert(GGML_OP_NONE);
-
-        for (int i = 0; i < GGML_OP_COUNT; ++i) {
-            const std::string key = std::string("GGML_OP_") + ggml_op_name(static_cast<ggml_op>(i));
-            if (table.count(key)) {
-                ops.insert(static_cast<ggml_op>(i));
-            }
-        }
-        for (int i = 0; i < GGML_UNARY_OP_COUNT; ++i) {
-            const std::string key = std::string("GGML_UNARY_OP_") + ggml_unary_op_name(static_cast<ggml_unary_op>(i));
-            if (table.count(key)) {
-                unary_ops.insert(static_cast<ggml_unary_op>(i));
-            }
-        }
-        for (int i = 0; i < GGML_GLU_OP_COUNT; ++i) {
-            const std::string key = std::string("GGML_GLU_OP_") + ggml_glu_op_name(static_cast<ggml_glu_op>(i));
-            if (table.count(key)) {
-                glu_ops.insert(static_cast<ggml_glu_op>(i));
-            }
-        }
-        return std::make_tuple(ops, unary_ops, glu_ops);
-    };
-    static const auto supported_sets = build_supported_sets();
-    static const auto & supported_ops = std::get<0>(supported_sets);
-    static const auto & supported_unary_ops = std::get<1>(supported_sets);
-    static const auto & supported_glu_ops = std::get<2>(supported_sets);
-
-    switch (op->op) {
-    case GGML_OP_UNARY: {
-        auto supported = supported_unary_ops.find(ggml_get_unary_op(op)) != supported_unary_ops.end();
-        if (!supported) {
-            // GGML_LOG_WARN("OpenVINO backend does not support unary op %s\n", ggml_unary_op_name(ggml_get_unary_op(op)));
-            return false;
-        }
-        if (ggml_get_unary_op(op) == GGML_UNARY_OP_EXP && op->type == GGML_TYPE_F32) {
-            return false;
-        }
-        break;
-    }
-    case GGML_OP_GLU: {
-        auto supported = supported_glu_ops.find(ggml_get_glu_op(op)) != supported_glu_ops.end();
-        if (!supported) {
-            // GGML_LOG_WARN("OpenVINO backend does not support GLU op %s\n", ggml_glu_op_name(ggml_get_glu_op(op)));
-            return false;
-        }
-        // if (has_view_op_input(op)) {
-        //     // GGML_LOG_WARN("OpenVINO backend does not support unary op %s with view input\n",
-        //     //               ggml_glu_op_name(ggml_get_glu_op(op)));
-        //     return false;
-        // }
-        if (op->src[1] == nullptr && op->src[0]->ne[0] % 2 != 0) {
-            // triggers bug in ov gpu
-            return false;
-        }
-        break;
-    }
-    default: {
-        auto supported = supported_ops.find(op->op) != supported_ops.end();
-        if (!supported) {
-            // GGML_LOG_WARN("OpenVINO backend does not support op %s\n", ggml_op_name(op->op));
-            return false;
-        }
-        static std::set<ggml_op> ops_not_support_view_input{};
-        if (ops_not_support_view_input.find(op->op) != ops_not_support_view_input.end() && has_view_op_input(op)) {
-            // GGML_LOG_WARN("OpenVINO backend does not support op %s with view input\n", ggml_op_name(op->op));
-            return false;
-        }
-    }
-    }
-
-    if (supported_types.find(op->type) == supported_types.end()) {
-        // GGML_LOG_WARN("OpenVINO backend does not support tensor type %s\n", ggml_type_name(op->type));
-        return false;
-    }
-    for (int i = 0; i < GGML_MAX_SRC; i++) {
-        auto * src = op->src[i];
-        if (src == nullptr) {
-            break;
-        }
-        if (supported_types.find(src->type) == supported_types.end()) {
-            // GGML_LOG_WARN("OpenVINO backend does not support tensor type %s\n", ggml_type_name(src->type));
-            return false;
-        }
-        const bool is_supported_3d_moe_expert =
-            op->op == GGML_OP_MUL_MAT_ID && i == 0 && (src->type == GGML_TYPE_MXFP4 || src->ne[3] == 1);
-        if (ggml_is_quantized(src->type) && src->ne[2] != 1 && !is_supported_3d_moe_expert) {
-            // GGML_LOG_WARN("OpenVINO backend does not support 3D quantized tensors\n");
-            return false;
-        }
-    }
-
-    if (is_op_unsupported_case(op)) {
-        return false;
-    }
-    return true;
+    return ggml_openvino_device_supports_op_impl(dev, op);
 }
 
 static bool ggml_backend_openvino_device_supports_buft(ggml_backend_dev_t dev, ggml_backend_buffer_type_t buft) {
